[CLN] website_sale: split product_attribute.py
