+++ addons/website_sale/models/__init__.py
@@ -15,6 +15,7 @@ from . import product_public_category
 from . import product_ribbon
 from . import product_tag
 from . import product_template
+from . import product_template_attribute_line
 from . import product_template_attribute_value
 from . import res_company
 from . import res_config_settings
